+++ BackEnd/pluto/src/main/java/com/subs/pluto/entity/UserNotification.java
@@ -1,6 +1,7 @@
 package com.subs.pluto.entity;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name="pl_user_notifications")
@@ -8,7 +9,7 @@ public class UserNotification {
     @Id
     @GeneratedValue
     @Column(name="user_notifications_id")
-    private Long id;
+    private BigDecimal id;
 
     @ManyToOne
     @JoinColumn(name = "user_id")
