+++ Fapi/fapi/fapi/src/main/java/com/subs/fapi/model/Company.java
@@ -1,9 +1,10 @@
 package com.subs.fapi.model;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class Company {
-    private Long id;
+    private BigDecimal id;
 
     private String name;
