+++ Fapi/fapi/fapi/src/main/java/com/subs/fapi/model/Subscription.java
@@ -2,10 +2,11 @@ package com.subs.fapi.model;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class Subscription {
-    private Long id;
+    private BigDecimal id;
 
     @JsonFormat(pattern = "yyyy-mm-dd")
     private Date subscriptionBuyingTime;
